+++ R/engine-r.R
@@ -1,42 +1,9 @@
-#' Bash-style brace expansion
-#'
-#' \code{expand_braces} performs brace expansions on strings,
-#' \code{str_expand_braces} is an alternate function that returns a list of character vectors.
-#' Made popular by Unix shells, brace expansion allows users to concisely generate
-#' certain character vectors by taking a single string and (recursively) expanding
-#' the comma-separated lists and double-period-separated integer and character
-#' sequences enclosed within braces in that string.
-#' The double-period-separated numeric integer expansion also supports padding the resulting numbers with zeros.
-#' @param string input character vector
-#' @return \code{expand_braces} returns a character vector while
-#'         \code{str_expand_braces} returns a list of character vectors.
-#'
-#' @examples
-#'   expand_braces("Foo{A..F}")
-#'   expand_braces("Foo{01..10}")
-#'   expand_braces("Foo{A..E..2}{1..5..2}")
-#'   expand_braces("Foo{-01..1}")
-#'   expand_braces("Foo{{d..d},{bar,biz}}.{py,bash}")
-#'   expand_braces(c("Foo{A..F}", "Bar.{py,bash}", "{{Biz}}"))
-#'   str_expand_braces(c("Foo{A..F}", "Bar.{py,bash}", "{{Biz}}"))
-#' @import stringr
-#' @export
-expand_braces <- function(string) {
-    c(str_expand_braces(string), recursive = TRUE)
-}
-
-#' @rdname expand_braces
-#' @export
-str_expand_braces <- function(string) {
-    lapply(string, expand_braces_helper)
-}
-
 brace_token <- "(?<!\\\\)\\{([^}]|\\\\\\})*(?<!\\\\)\\}"
 has_brace <- function(string) {
     str_detect(string, brace_token)
 }
 
-expand_braces_helper <- function(string, process = TRUE) {
+expand_braces_r <- function(string, process = TRUE) {
     locations <- get_locations(string) # Find brace starts and ends
     n <- nrow(locations)
     if (n == 0) return(process_string(string))
@@ -117,7 +84,7 @@ expand_brace <- function(string, locations, i) {
     } else if (has_periods(inner)) {
         expand_periods(inner)
     } else if (has_brace(inner)) {
-        paste0("{", expand_braces_helper(inner, FALSE), "}")
+        paste0("{", expand_braces_r(inner, FALSE), "}")
     } else {
         brace
     }
@@ -131,7 +98,7 @@ has_comma <- function(string) {
 
 expand_comma <- function(string) {
     elements <- str_split(string, comma_token)[[1]]
-    elements <- lapply(elements, expand_braces_helper, FALSE)
+    elements <- lapply(elements, expand_braces_r, FALSE)
     elements <- c(elements, recursive = TRUE)
     elements
 }
@@ -184,8 +151,6 @@ expand_periods <- function(string) {
         } else {
             values
         }
-    } else {
-        paste0("{", string, "}")
     }
 }